javascript sdk wrapper for generated code, next rev.
