+++ examples/connect-example.js
@@ -1,20 +1,8 @@
-let AmberApiServer = require('amber-javascript-sdk');
+let MyClient = require('amber-javascript-sdk')
 
-// create oauth2 token using credentials
-let apiInstance = new AmberApiServer.DefaultApi();
+// create amber instance
+let amberInstance = new MyClient.AmberClient()
 
-// let body = new AmberApiServer.PostAuth2Request();
-username = "your-username";
-password = "your-password";
-
-let defaultClient = AmberApiServer.ApiClient.instance;
-let authorize_amber_pool = defaultClient.authentications['authorize-amber-pool'];
-
-let body = new AmberApiServer.PostAuth2Request(password, username);
-apiInstance.postOauth2(body, (error, data, response) => {
-    if (error) {
-        console.log(error)
-    } else {
-        console.log(data)
-    }
-});
+amberInstance.getSensorsRequest().then(function (data) {
+    console.log("getSensorsResponse: %o", data)
+})
