+++ package.json
@@ -5,7 +5,7 @@
   "license": "Unlicense",
   "main": "dist/sdk.js",
   "scripts": {
-    "package": "rm -rf dist && mkdir dist && babel sdk.js ./src -d dist --copy-files",
+    "package": "rm -rf dist && mkdir dist && babel ./src -d dist --copy-files",
     "test": "mocha --compilers js:babel-core/register --recursive",
     "generate": "cp package.json package.json.backup && swagger-codegen generate -DmoduleName=amber-javascript-sdk -DpackageName=amber-javascript-sdk -DprojectDescription='Boon_Logic_Amber_SDK' -i swagger.json -l javascript -o . && mv package.json.backup package.json"
   },
